Moved observing proximity events inside HomeActivity
