+++ app/src/main/java/net/squanchy/favorites/FavoritesComponent.java
@@ -7,7 +7,6 @@ import net.squanchy.injection.ActivityLifecycle;
 import net.squanchy.injection.ApplicationComponent;
 import net.squanchy.navigation.NavigationModule;
 import net.squanchy.navigation.Navigator;
-import net.squanchy.service.proximity.injection.ProximityService;
 
 import dagger.Component;
 
@@ -20,8 +19,6 @@ interface FavoritesComponent {
     Context context();
 
     Navigator navigator();
-
-    ProximityService proxService();
 
     Analytics analytics();
 }
