+++ app/src/main/java/net/squanchy/schedule/ScheduleComponent.java
@@ -7,21 +7,18 @@ import net.squanchy.injection.ActivityLifecycle;
 import net.squanchy.injection.ApplicationComponent;
 import net.squanchy.navigation.NavigationModule;
 import net.squanchy.navigation.Navigator;
-import net.squanchy.service.proximity.injection.ProximityService;
 
 import dagger.Component;
 
 @ActivityLifecycle
 @Component(modules = {ScheduleModule.class, NavigationModule.class}, dependencies = ApplicationComponent.class)
-public interface ScheduleComponent {
+interface ScheduleComponent {
 
     ScheduleService service();
 
     Context context();
 
     Navigator navigator();
-
-    ProximityService proxService();
 
     Analytics analytics();
 }
