+++ app/src/main/java/net/squanchy/schedule/SchedulePageView.java
@@ -12,18 +12,17 @@ import android.support.v7.widget.Toolbar;
 import android.text.Spanned;
 import android.util.AttributeSet;
 import android.view.View;
+
 import net.squanchy.R;
 import net.squanchy.analytics.Analytics;
 import net.squanchy.analytics.ContentType;
 import net.squanchy.navigation.Navigator;
-import net.squanchy.proximity.ProximityEvent;
 import net.squanchy.schedule.domain.view.Event;
 import net.squanchy.schedule.domain.view.Schedule;
 import net.squanchy.schedule.view.ScheduleViewPagerAdapter;
-import net.squanchy.service.proximity.injection.ProximityService;
 
 import io.reactivex.android.schedulers.AndroidSchedulers;
-import io.reactivex.disposables.CompositeDisposable;
+import io.reactivex.disposables.Disposable;
 import uk.co.chrisjenx.calligraphy.CalligraphyTypefaceSpan;
 import uk.co.chrisjenx.calligraphy.CalligraphyUtils;
 import uk.co.chrisjenx.calligraphy.TypefaceUtils;
@@ -32,10 +31,9 @@ public class SchedulePageView extends CoordinatorLayout {
 
     private ScheduleViewPagerAdapter viewPagerAdapter;
     private View progressBar;
-    private CompositeDisposable subscriptions;
+    private Disposable subscription;
     private ScheduleService service;
     private Navigator navigate;
-    private ProximityService proximityService;
     private Analytics analytics;
 
     public SchedulePageView(Context context, AttributeSet attrs) {
@@ -55,7 +53,6 @@ public class SchedulePageView extends CoordinatorLayout {
         service = component.service();
         navigate = component.navigator();
         analytics = component.analytics();
-        proximityService = component.proxService();
 
         progressBar = findViewById(R.id.progressbar);
 
@@ -101,16 +98,9 @@ public class SchedulePageView extends CoordinatorLayout {
     @Override
     public void onAttachedToWindow() {
         super.onAttachedToWindow();
-        subscriptions = new CompositeDisposable();
-        subscriptions.add(
-                service.schedule()
-                        .observeOn(AndroidSchedulers.mainThread())
-                        .subscribe(schedule -> updateWith(schedule, this::onEventClicked)));
-
-        subscriptions.add(
-                proximityService.observeProximityEvents()
-                        .observeOn(AndroidSchedulers.mainThread())
-                        .subscribe(this::handleProximityEvent));
+        subscription = service.schedule()
+                .observeOn(AndroidSchedulers.mainThread())
+                .subscribe(schedule -> updateWith(schedule, this::onEventClicked));
     }
 
     private void onEventClicked(Event event) {
@@ -118,14 +108,10 @@ public class SchedulePageView extends CoordinatorLayout {
         navigate.toEventDetails(event.id());
     }
 
-    private void handleProximityEvent(ProximityEvent proximityEvent) {
-        // TODO do something with the event, like showing feedback or opening an event detail
-    }
-
     @Override
     public void onDetachedFromWindow() {
         super.onDetachedFromWindow();
-        subscriptions.dispose();
+        subscription.dispose();
     }
 
     private void hackToApplyTypefaces(TabLayout tabLayout) {
